# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.ChangePasswordRequest:
    properties:
      new_password:
        type: string
      old_password:
        type: string
      user_id:
        type: string
    type: object
...
  dto.CreateUserRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
...
  dto.ListUsersResponse:
    properties:
      page:
        $ref: '#/definitions/dto.Page'
      result:
        items:
          $ref: '#/definitions/dto.UserResponse'
        type: array
    type: object
  dto.Page:
    properties:
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
...
  dto.UserResponse:
    properties:
      created_at:
        type: string
      expired_at:
        type: string
      id:
        type: string
      locked:
        type: boolean
      username:
        type: string
    type: object
...
  jsonutil.Message:
    properties:
      code:
        type: integer
      detail: {}
      message:
        type: string
    type: object
info:
  contact: {}
  description: Stash Password Manager API Server
  title: Stash API Server
  version: "0.0"
paths:
  /auth/change-password:
    patch:
      consumes:
      - application/json
      description: change password for current user or (if admin) for specified user
      parameters:
      - description: change password request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/jsonutil.Message'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/jsonutil.Message'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/jsonutil.Message'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/jsonutil.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/jsonutil.Message'
      security:
      - BearerAuth: []
      summary: Change Password
      tags:
      - Accounts
...
  /auth/whoami:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/jsonutil.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/jsonutil.Message'
      security:
      - BearerAuth: []
      summary: Get current user information
      tags:
      - Accounts
...
  /users:
    get:
      consumes:
      - application/json
      parameters:
      - description: limit
        in: query
        minimum: 0
        name: limit
        type: integer
      - description: offset
        in: query
        minimum: 0
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ListUsersResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/jsonutil.Message'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/jsonutil.Message'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/jsonutil.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/jsonutil.Message'
      security:
      - BearerAuth: []
      summary: List users
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Requires admin role.
      parameters:
      - description: user credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/jsonutil.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/jsonutil.Message'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/jsonutil.Message'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/jsonutil.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/jsonutil.Message'
      security:
      - BearerAuth: []
      summary: Create User
      tags:
      - Users
  /users/{username_or_id}:
    get:
      consumes:
      - application/json
      description: Returns user by id(uuid) or (if request param is not uuid) by username
      parameters:
      - description: Username or ID
        in: path
        name: username_or_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/jsonutil.Message'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/jsonutil.Message'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/jsonutil.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/jsonutil.Message'
      security:
      - BearerAuth: []
      summary: Get username by username or id
      tags:
      - Users
securityDefinitions:
  BearerAuth:
    description: 'Bearer token. Format: "Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"